Extend to a maximum of 256 component types (#313)
Rework `Mask` to hold an array of 4 x `uint64` to represent 256 bits. Everything else comes into place almost by itself.

Alternative to #312:

- `Filter.Match` takes a mask pointer as argument
- `Mask` is copied when using as a filter (is pointer receiver in #312)

Primarily avoids `All(ids...)` either returning a pointer, or requiring users to store the result (temporarily) to be able to get a pointer to it.

Performance remains unchanged, except that `MaskFilter` matches are about 3 times slower (i.e. filters with component inclusion and exclusion). Not sure why, as the mask functions used there are as fast as before.

## Commits

* Extend to a maximum of 256 component types
* fix IdMap for 265 possible IDs
* filters match against pointer to mask, to avoid copying
* minor optimization: check for zero mask at end of MaskFilter conditions
* make internal Mask representation private
* update changelog and readme